requirements-streamlit.txt
streamlit>=1.32
pandas>=2.1
numpy>=1.26
python-dotenv>=1.0
openai>=1.40
